pyproject.toml
[tool.poetry]
name = "projeto-loja"
version = "0.1.0"
description = ""
authors = ["Guilherme <gui.santos.sa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.10.6"
sqlalchemy = "^2.0.37"
faker = "^35.2.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
streamlit = "^1.41.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
